+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/__init__.py
@@ -40,7 +40,8 @@ from ._deployment.resource_requirements_settings import ResourceRequirementsSett
 from ._deployment.scale_settings import DefaultScaleSettings, TargetUtilizationScaleSettings
 from ._endpoint.batch_endpoint import BatchEndpoint
 from ._endpoint.endpoint import Endpoint
-from ._endpoint.online_endpoint import KubernetesOnlineEndpoint, ManagedOnlineEndpoint, OnlineEndpoint
+from ._endpoint.online_endpoint import KubernetesOnlineEndpoint, ManagedOnlineEndpoint, OnlineEndpoint,\
+    EndpointAuthKeys, EndpointAuthToken
 from ._job.command_job import CommandJob
 from ._job.compute_configuration import ComputeConfiguration
 from ._job.input_port import InputPort
@@ -84,8 +85,6 @@ from ._system_data import SystemData
 from ._validation import ValidationResult
 from ._workspace.connections.workspace_connection import WorkspaceConnection
 from ._workspace.customer_managed_key import CustomerManagedKey
-from ._workspace.identity import ManagedServiceIdentity
-from ._workspace.identity import UserAssignedIdentity as WorkspaceUserAssignedIdentity
 from ._workspace.private_endpoint import EndpointConnection, PrivateEndpoint
 from ._workspace.workspace import Workspace
 from ._workspace.workspace_keys import WorkspaceKeys, NotebookAccessKeys, ContainerRegistryCredential
@@ -145,8 +144,6 @@ __all__ = [
     "Workspace",
     "WorkspaceKeys",
     "WorkspaceConnection",
-    "ManagedServiceIdentity",
-    "WorkspaceUserAssignedIdentity", # pylint: disable=naming-mismatch
     "PrivateEndpoint",
     "EndpointConnection",
     "CustomerManagedKey",
@@ -222,5 +219,7 @@ __all__ = [
     "AmlTokenConfiguration",
     "IdentityConfiguration",
     "NotebookAccessKeys",
-    "ContainerRegistryCredential"
+    "ContainerRegistryCredential",
+    "EndpointAuthKeys",
+    "EndpointAuthToken"
 ]