+++ sdk/ml/azure-ai-ml/tests/automl_job/unittests/test_text_ner_job.py
@@ -1,8 +1,8 @@
 import pytest
 
-from azure.ai.ml import UserIdentity
-from azure.ai.ml._restclient.v2022_06_01_preview.models import AutoMLJob as RestAutoMLJob
-from azure.ai.ml._restclient.v2022_06_01_preview.models import (
+from azure.ai.ml import UserIdentityConfiguration
+from azure.ai.ml._restclient.v2022_10_01_preview.models import AutoMLJob as RestAutoMLJob
+from azure.ai.ml._restclient.v2022_10_01_preview.models import (
     JobBase,
     LogVerbosity,
     MLTableJobInput,
@@ -11,8 +11,8 @@ from azure.ai.ml._restclient.v2022_06_01_preview.models import (
     TaskType,
     TextNer,
 )
-from azure.ai.ml._restclient.v2022_06_01_preview.models import UserIdentity as RestUserIdentity
-from azure.ai.ml._restclient.v2022_06_01_preview.models._azure_machine_learning_workspaces_enums import (
+from azure.ai.ml._restclient.v2022_10_01_preview.models import UserIdentity as RestUserIdentity
+from azure.ai.ml._restclient.v2022_10_01_preview.models._azure_machine_learning_workspaces_enums import (
     ClassificationPrimaryMetrics,
 )
 from azure.ai.ml._utils.utils import to_iso_duration_format_mins
@@ -96,7 +96,7 @@ class TestAutoMLTextNerJob:
         timeout = 30
         dataset_language = "eng"
 
-        identity = UserIdentity()
+        identity = UserIdentityConfiguration()
         job = text_ner(
             training_data=Input(type=AssetTypes.MLTABLE, path=training_data_uri),
             validation_data=Input(type=AssetTypes.MLTABLE, path=validation_data_uri),
@@ -141,7 +141,7 @@ class TestAutoMLTextNerJob:
         timeout = 30
         dataset_language = "eng"
 
-        identity = UserIdentity()
+        identity = UserIdentityConfiguration()
         expected_job = text_ner(
             training_data=Input(type=AssetTypes.MLTABLE, path=training_data_uri),
             validation_data=Input(type=AssetTypes.MLTABLE, path=validation_data_uri),
@@ -165,7 +165,7 @@ class TestAutoMLTextNerJob:
             ),
             featurization_settings=NlpVerticalFeaturizationSettings(dataset_language=dataset_language),
         )
-        job_data = JobBase(properties=RestAutoMLJob(task_details=task_details, identity=identity._to_rest_object()))
+        job_data = JobBase(properties=RestAutoMLJob(task_details=task_details, identity=identity._to_job_rest_object()))
         # Test converting REST object to Job
         converted_to_job = TextNerJob._from_rest_object(job_data)
         assert isinstance(converted_to_job, TextNerJob)